% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_boxplot.R
\name{nba_boxplot}
\alias{nba_boxplot}
\title{Creating Boxplot for NBA Stats}
\usage{
nba_boxplot(dataset, yaxis, xaxis)
}
\arguments{
\item{dataset}{dataframe
This dataframe is created after using the nba_scraper.R function or if
the csv has already been loaded, read the csv in and pass it as the parameter.}

\item{yaxis}{string
The parameter of interest examples: Points, 3_Pointers, Turnovers}

\item{xaxis}{string
The categorical variable of interest examples: Team, Position}
}
\value{
ggplot boxplot
}
\description{
Creates a boxplot of the categorical variable of interest on the y-axis and
 the stat of interest on the x-axis.
}
\examples{
nba_boxplot(NBA_reg_01_02, Team, Points)
}
