% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_scraper_pseudo.R
\name{nba_scraper_pseudo}
\alias{nba_scraper_pseudo}
\title{Scrape NBA statistics data from ESPN into CSV file}
\usage{
nba_scraper_pseudo(
  season_year = 2018,
  season_type = "regular",
  port = 4445L,
  sel_browser = "chrome",
  csv_path = NULL
)
}
\arguments{
\item{season_year}{int from 2001 to 2019 (upper limit based on latest year)}

\item{season_type}{string. Either "regular" or "postseason"}

\item{port}{int with L suffix. Must not be negative.}

\item{sel_browser}{string. Either "chrome" or "firefox".}

\item{csv_path}{string for csv file. Must end with ".csv".}
}
\value{
A tibble of scraped ESPN NBA data
}
\description{
This function is a pseudo version of the rsketball::nba_scraper function and will be
used for the grading rubric of DSCI 524 with respect to testing completion. Reason is
because the original rsketball::nba_scraper function can only work with a running
Docker container that represents the driver for RSelenium, which renders it as
untestable when using devtools::check(). Thus, this pseudo function comprises of
most of the non-Selenium related scraping code, and serves as a proxy for testing
a large portion of the original rsketball::nba_scraper function code.
}
\details{
Scrape the tabular data from ESPN NBA website using RSelenium and returns a tibble
of the data. Users can specify the seaason year and season type. User should also
specify the port for Selenium driver, and the driver type. By default, the function
will not write to csv until an input for "csv_path" is given.
}
