% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_team_stats.R
\name{nba_team_stats}
\alias{nba_team_stats}
\title{Generate summary stats for NBA players}
\usage{
nba_team_stats(
  nba_data,
  stats_filter = c(),
  teams_filter = c(),
  positions_filter = c()
)
}
\arguments{
\item{nba_data}{tibble of scraped ESPN NBA data}

\item{stats_filter}{character vector}

\item{teams_filter}{character vector}

\item{positions_filter}{character vector}
}
\value{
A tibble
}
\description{
The function provides descriptive (mean, median, 25%, and 75% quantiles) team statistics of NBA data. Users can specify which
statistic of interest (3PA, 3PM, etc) along with teams of interest (GS, HOU, etc). If positions of interest (C, PG, etc)
are specified, the returned tibble depicts relevant descriptive statistics for the relevant positions in the relevant teams.
}
\examples{
\donttest{
nba_data = read_csv(
"https://raw.githubusercontent.com/kfoofw/nba_espn/master/0.data/NBA_reg_2018-2019.csv"
)
# Find descriptive stats for all teams without position
nba_team_stats(nba_data)

# Find specific stats (3PM, 3PA) for specific teams (GS, HOU) for specific positions (PG, C)
nba_team_stats(nba_data, stats_filter = c("3PM","3PA"), teams_filter = c("GSW","HOU"), positions_filter = c("C","PG"))

# Find specific stats (3PM, 3PA) for specific teams (GS, HOU) without positions_filter inputs
nba_team_stats(nba_data, stats_filter = c("3PM","3PA"), teams_filter = c("GS","HOU"))

# Find specific stats (3PM, 3PA) for all teams (unspecified) for specific positions (PG, C)
nba_team_stats(nba_data, stats_filter = c("3PM","3PA"), positions_filter = c("C","PG"))
}
}
