% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_boxplot.R
\name{nba_boxplot}
\alias{nba_boxplot}
\title{Creating Boxplot for NBA Stats}
\usage{
nba_boxplot(dataset, position = NULL, teams = c(NULL), stats)
}
\arguments{
\item{dataset}{dataframe
This dataframe is created after using the nba_scraper.R function or if
the csv has already been loaded, read the csv in and pass it as the parameter.}

\item{position}{string
To specify as "POS" if you want to look at Position boxplot.}

\item{teams}{string
The names of teams you want to create boxplot for.
*Note you can only choose one of position or teams as a variable to vizualize}

\item{stats}{string
The numeric variable of interest examples:  Points, 3_Pointers, Turnovers}
}
\value{
ggplot boxplot
}
\description{
Creates a boxplot of the categorical variable of interest on the y-axis and
 the stat of interest on the x-axis.
}
\examples{
nba_2018 <- data.frame(POS= c("C", "FOR", "PO","FOR", "C"), Team = c("ORL", "UTAH", "LAC", "MIN", "BOS"), GP = c(3, 5, 5, 2, 1))
nba_boxplot(nba_2018, position= "POS", teams= NULL, stats= "GP")

}
