% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_ranking.R
\name{nba_ranking}
\alias{nba_ranking}
\title{Generates a ranking and a visualization based on a column of a dataset}
\usage{
nba_ranking(data, column, by, top)
}
\arguments{
\item{data}{The dataframe to make the ranking from}

\item{column}{The column from the dataset to rank}

\item{by}{The column from the dataset to rank by}

\item{top}{The number of elements in the ranking}
}
\value{
ggplot visualization with the ranking
}
\description{
This function generates a ranking and a visualization for a variable, according to a different variable the user sets.
}
\examples{
nba_ranking(nba_data, 'Player', 'TD3', 10)
}
