% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_scraper.R
\name{nba_scraper}
\alias{nba_scraper}
\title{Scrape NBA statistics data from ESPN into CSV file}
\usage{
nba_scraper(season_year, season_type, csv_path, nba_data_env_name = NULL)
}
\arguments{
\item{season_year}{string}

\item{season_type}{string}

\item{csv_path}{string}

\item{nba_data_env_name}{string}
}
\description{
Scrape the tabular data from ESPN NBA website into a csv file.
}
\examples{
# Regular season for 2018/2019 while assigning it into global env as an object called nba_2018
nba_scraper(season_year = "2018", season_type = "regular", csv_path = "~/nba_scraped_data/nba_2018_reg.csv",  nba_data_env_name = "nba_2018")

# Playoff season for 2016/2017 without assigning it into global env as an object called nba_2018
nba_scraper(season_year = "2016", season_type = "playoff", csv_path = "~/nba_scraped_data/nba_2018_reg.csv",  nba_data_env_name = NULL)
}
