% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_team_stats.R
\name{nba_team_stats}
\alias{nba_team_stats}
\title{Generate summary stats for NBA players}
\usage{
nba_team_stats(
  nba_data,
  stats_filter = c(),
  teams_filter = c(),
  positions_filter = c()
)
}
\arguments{
\item{nba_data}{tibble of scraped ESPN NBA data}

\item{stats_filter}{character vector}

\item{teams_filter}{character vector}

\item{positions_filter}{character vector}
}
\value{
A tibble
}
\description{
The function provides descriptive (mean, median, 25%, and 75% quantiles) team statistics of NBA data. Users can specify which
statistic of interest (3PA, 3PM, etc) along with teams of interest (GS, HOU, etc). If positions of interest (C, PG, etc)
are specified, the returned tibble depicts relevant descriptive statistics for the relevant positions in the relevant teams.

For reference on the scraped data columns information, please refer to the
dataset description:
https://github.com/UBC-MDS/rsketball/blob/master/dataset_description.md

For detailed use cases, please refer to the vignette:
https://ubc-mds.github.io/rsketball/articles/rsketball-vignette.html
}
\examples{
nba_data <- tibble::tibble(NAME = c("James", "Steph", "Bosh", "Klay", "Kobe"),
                          TEAM = c("MIA","MIA","MIA","GS","GS"),
                          POS = c("SF", "PG", "C", "C", "PG"),
                          PTS = c(5,4,3,2,10),
                          TO = c(1,2,3,4,3))

# Find descriptive stats for all teams without position
nba_team_stats(nba_data)

# Find specific stats (PTS, TO) for specific teams (GS, MIA) for specific positions (PG, C)
nba_team_stats(nba_data, stats_filter = c("PTS","TO"),
               teams_filter = c("GS","MIA"), positions_filter = c("C","PG"))

# Find specific stats (PTS, TO) for specific teams (GS) without positions_filter inputs
nba_team_stats(nba_data, stats_filter = c("PTS","TO"), teams_filter = c("GS"))

# Find specific stats (PTS, TO) for all individual teams (unspecified)
# for specific positions (PG, C)
nba_team_stats(nba_data, stats_filter = c("PTS","TO"), positions_filter = c("PG"))
}
